API: move recover_threshold to its own file, there's a lot more to do and it needs its own space
