+++ tests/test_plan_stubs.py
@@ -8,7 +8,7 @@ import logging
 from bluesky.plans import run_wrapper, scan
 
 from pswalker.plan_stubs import (prep_img_motors, as_list, verify_all,
-                                 match_condition, recover_threshold,
+                                 match_condition,
                                  slit_scan_area_comp)
 from pswalker.utils.exceptions import RecoverDone, RecoverFail
 from .utils import plan_stash, SlowSoftPositioner, MotorSignal, collector
@@ -191,71 +191,6 @@ def test_match_condition_timeout(RE, mot_and_sig):
     assert mot.position < 5
     # If the motor did not reach 5, we timed out
 
-
-@pytest.mark.timeout(tmo)
-def test_recover_threshold_success(RE, mot_and_sig):
-    logger.debug("test_recover_threshold_success")
-    mot, sig = mot_and_sig
-    with pytest.raises(RecoverDone):
-        RE(run_wrapper(recover_threshold(sig, 20, mot, +1)))
-    assert mot.position < 21
-    # If we stopped right after 20, we recovered
-
-
-@pytest.mark.timeout(tmo)
-def test_recover_threshold_success_no_stop(RE, mot_and_sig):
-    logger.debug("test_recover_threshold_success_no_stop")
-    mot, sig = mot_and_sig
-    mot.delay = 0
-    with pytest.raises(RecoverDone):
-        RE(run_wrapper(recover_threshold(sig, 20, mot, +1, has_stop=False)))
-    assert 59 < mot.position < 61
-    # If we went halfway between 20 and 100, it worked
-
-
-@pytest.mark.timeout(tmo)
-def test_recover_threshold_success_reverse(RE, mot_and_sig):
-    logger.debug("test_recover_threshold_success_reverse")
-    mot, sig = mot_and_sig
-    with pytest.raises(RecoverDone):
-        RE(run_wrapper(recover_threshold(sig, -1, mot, +1)))
-    assert mot.position > -2
-    # If we stopped right after -1, we recovered
-
-
-@pytest.mark.timeout(tmo)
-def test_recover_threshold_failure(RE, mot_and_sig):
-    logger.debug("test_recover_threshold_failure")
-    mot, sig = mot_and_sig
-    with pytest.raises(RecoverFail):
-        RE(run_wrapper(recover_threshold(sig, 101, mot, +1)))
-    assert mot.position == -100
-    # We got to the end of the negative direction, we failed
-
-
-@pytest.mark.timeout(tmo)
-def test_recover_threshold_failure_no_stop(RE, mot_and_sig):
-    logger.debug("test_recover_threshold_failure_no_stop")
-    mot, sig = mot_and_sig
-    mot.delay = 0
-    with pytest.raises(RecoverFail):
-        RE(run_wrapper(recover_threshold(sig, 101, mot, +1, has_stop=False)))
-    assert mot.position == -100
-    # We got to the end of the negative direction, we failed
-
-
-@pytest.mark.timeout(tmo)
-def test_recover_threshold_timeout_failure(RE, mot_and_sig):
-    logger.debug("test_recover_threshold_timeout_failure")
-    mot, sig = mot_and_sig
-    # Make the motor slower to guarantee a timeout
-    mot.n_steps = 5000
-    with pytest.raises(RecoverFail):
-        RE(run_wrapper(recover_threshold(sig, 50, mot, +1, timeout=0.1)))
-    pos = mot.position
-    assert not 49 < pos < 51
-    assert mot.position not in (100, -100)
-    # If we didn't reach the goal or either end, we timed out
 
 @pytest.mark.timeout(tmo)
 def test_slit_scan_area_compare(RE):
